move authentication to layout
