+++ app-web/package.json
@@ -100,6 +100,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^21.2.1",
     "jest-localstorage-mock": "^2.2.0",
-    "prettier": "1.7.4"
+    "prettier": "1.7.4",
+    "redux-mock-store": "^1.5.3"
   }
 }
